refactor: 폴더 구성 변경
