+++ src/main/java/com/programmers/springbootjpa/domain/customer/Address.java
@@ -1,4 +1,4 @@
-package com.programmers.springbootjpa.domain;
+package com.programmers.springbootjpa.domain.customer;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
