+++ src/main/java/com/programmers/springbootjpa/repository/CustomerRepository.java
@@ -1,6 +1,6 @@
 package com.programmers.springbootjpa.repository;
 
-import com.programmers.springbootjpa.domain.Customer;
+import com.programmers.springbootjpa.domain.customer.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository extends JpaRepository<Customer, Long> {
